fameium/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.1.0
awsebcli==3.18.2
bcrypt==3.2.0
black==20.8b1
blessed==1.17.10
botocore==1.15.49
CacheControl==0.12.6
cached-property==1.5.1
cachy==0.3.0
cement==2.8.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cleo==0.7.6
click==7.1.2
clikit==0.4.3
colorama==0.4.3
cryptography==3.1
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
html5lib==1.1
idna==2.10
jmespath==0.10.0
jsonschema==3.2.0
keyring==20.0.1
lockfile==0.12.2
msgpack==1.0.0
mypy-extensions==0.4.3
paramiko==2.7.2
pastel==0.2.0
pathspec==0.8.0
pexpect==4.8.0
pkginfo==1.5.0.1
poetry==1.0.10
psycopg2==2.8.5
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.20
pylev==1.3.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-toolbelt==0.8.0
semantic-version==2.5.0
shellingham==1.3.2
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.3
toml==0.10.1
tomlkit==0.5.11
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
